function that initializes a variable of type struct dog

+++ 0x0E-structures_typedef/1-init_dog.c
@@ -9,7 +9,7 @@
  * @owner:owner to initialize
  */
 
-void init_dog(struct dog *d, char *name, double age, char *owner)
+void init_dog(struct dog *d, char *name, float age, char *owner)
 {
 	if (d == NULL)
 	{
